app/src/main/java/com/abduldev/artbookapptestingguide/db/dao/ArtDao.kt
package com.abduldev.artbookapptestingguide.db.dao

import androidx.lifecycle.LiveData
import androidx.room.Dao
import androidx.room.Delete
import androidx.room.Insert
import androidx.room.OnConflictStrategy
import androidx.room.Query
import com.abduldev.artbookapptestingguide.db.enities.Art


@Dao
interface ArtDao {

    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insertAtr(art: Art)

    @Delete
    suspend fun deleteArt(art: Art)

    @Query("SELECT * FROM arts")
    fun getAllArts(): LiveData<List<Art>>

}